Update FS_model.gaml
The carrying capacity was changed to use the mortality based on the number of pupae, tenoral and emerged there were from the larval tree. Another change is that the nb_cohort is updated to reflect the max capacity for the tree that it is within (so the nb_cohort is reduced. Dramatically in some cases)