Update FSFM_v2.gaml
the larvae_exited fruit is replacing teneral + pupal as this was incorrect as these values changed overtime. The value of larvae_exited_fruit only increases when larvae exit the fruit (so at the end of the larval state). This determines the carrying capacity of the whole tree. The max_larvae_per_fruit is set very low for testing purposes. UNDO the global save file.