Protein bait spray included
Protein bait spray is now included and works. There is an issue of the type of death not able to be recorded in the global. I tried to add protien_spray <- protein_spray + 1; to the global when the fly dies. Didn't work. Also tried to tag the fly with a boolean when sprayed so they can be killed the next day, again this did not work. As it is the model works in this form and information on how the fly died can be counted from the fly file.