Files used in simulations for FFHF model
These files are required for the simulations used in the Fruit Fly Host Foraging model. Each ASCII (.asc) file is a 50 x 50 grid landscape with each cell size 5 x 5m. The "proba_of_staying.csv" file is the probability matrix of a fly staying in an affiliated tree depending on the day since exposure.